Change term "виринання" to term "підйом"
I believe that term "виринання" is a better translation for the "hoisting" term (as well in the Russian tech literature the term "поднятие" is used. Term "виринання" can be used to describe the behavior of the event when they bubble through the elements, that have a parent-child relationship. To use this term here can confused, as variables do not bubble through all the function (the same as events do), but just come to the beginning of the function